requirements.txt
asgiref==3.9.0
decouple==0.0.7
Django==5.2.4
django-dotenv==1.4.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
mysqlclient==2.2.7
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.14.1
